Use more size_t instead of int internally

+++ Game.c
@@ -17,7 +17,7 @@ void disposeGame(Game* game) {
 }
 
 int getTurnNumber(Game* game) {
-    return (int)game->currentTurn;
+    return game->currentTurn;
 }
 
 PlayerId getWhoseTurn(Game* game) {
